+++ src/main/java/com/greenfox/lvlup/model/dto/user/UserBadgeDTO.java
@@ -1,16 +1,20 @@
-package com.greenfox.lvlup.model.dto;
+package com.greenfox.lvlup.model.dto.user;
 
-public class BadgeLevelForArchetypeDTO {
-
+public class UserBadgeDTO {
   private String name;
   private int level;
 
-  public BadgeLevelForArchetypeDTO() {
+  public UserBadgeDTO(String name) {
+    this.name = name;
+    this.level = 1;
   }
 
-  public BadgeLevelForArchetypeDTO(String name, int level) {
+  public UserBadgeDTO(String name, int level) {
     this.name = name;
     this.level = level;
+  }
+
+  public UserBadgeDTO() {
   }
 
   public String getName() {
